+++ app/src/main/java/com/r0adkll/deckbuilder/arch/data/features/offline/repository/DefaultOfflineRepository.kt
@@ -4,7 +4,9 @@ import android.content.Context
 import com.jakewharton.rxrelay2.BehaviorRelay
 import com.jakewharton.rxrelay2.Relay
 import com.r0adkll.deckbuilder.arch.data.AppPreferences
+import com.r0adkll.deckbuilder.arch.data.features.offline.cache.OfflineCache
 import com.r0adkll.deckbuilder.arch.data.features.offline.service.CacheService
+import com.r0adkll.deckbuilder.arch.domain.features.expansions.model.Expansion
 import com.r0adkll.deckbuilder.arch.domain.features.offline.model.CacheStatus
 import com.r0adkll.deckbuilder.arch.domain.features.offline.model.DownloadRequest
 import com.r0adkll.deckbuilder.arch.domain.features.offline.model.OfflineStatus
@@ -12,12 +14,15 @@ import com.r0adkll.deckbuilder.arch.domain.features.offline.repository.OfflineRe
 import com.r0adkll.deckbuilder.internal.di.scopes.AppScope
 import io.reactivex.Observable
 import io.reactivex.android.schedulers.AndroidSchedulers
+import io.reactivex.disposables.Disposable
+import timber.log.Timber
 import javax.inject.Inject
 
 @AppScope
 class DefaultOfflineRepository @Inject constructor(
     val context: Context,
-    val preferences: AppPreferences
+    val preferences: AppPreferences,
+    val cache: OfflineCache
 ) : OfflineRepository, OfflineStatusConsumer {
 
     override var status = OfflineStatus()
@@ -26,31 +31,44 @@ class DefaultOfflineRepository @Inject constructor(
             statusRelay.accept(value)
         }
 
+    private var offlineStatusDisposable: Disposable? = null
     private val statusRelay: Relay<OfflineStatus> = BehaviorRelay.create<OfflineStatus>().toSerialized()
 
     init {
-        // Load saved preferences
-        val offlineStatus = HashMap<String, CacheStatus>()
-        preferences.offlineExpansions.get().forEach {
-            offlineStatus[it] = CacheStatus.Cached
-        }
-        status = OfflineStatus(offlineStatus)
+        offlineStatusDisposable = cache.getOfflineStatus()
+            .subscribe({
+                status = it
+            }, {
+                Timber.w(it, "Unable to get the current cache status")
+            })
     }
 
     override fun download(request: DownloadRequest) {
         // Filter any expansions that we have marked as cached
-        val cached = preferences.offlineExpansions.get()
-        val filteredExpansions = request.expansion.filter { !cached.contains(it.code) }
+        val filteredExpansions = request.expansion.filter {
+            val status = status.expansions[it]
+            status == null || status is CacheStatus.Empty
+        }
 
         // Flag all expansions in request
         var statusChanges = status
         filteredExpansions.forEach {
-            statusChanges = statusChanges.set(it.code to CacheStatus.Queued)
+            statusChanges = statusChanges.set(it to CacheStatus.Queued)
         }
         status = statusChanges
 
-        // Start service
-        CacheService.start(context, DownloadRequest(filteredExpansions, request.downloadImages))
+        CacheService.start(context, request.copy(expansion = filteredExpansions))
+    }
+
+    override fun delete(expansion: Expansion?): Observable<Unit> {
+        return cache.delete(expansion)
+            .flatMap {
+                cache.getOfflineStatus()
+                    .doOnNext {
+                        status = it
+                    }
+                    .map { Unit }
+            }
     }
 
     override fun observeStatus(): Observable<OfflineStatus> {