+++ app/src/main/java/com/r0adkll/deckbuilder/arch/ui/features/carddetail/CardDetailActivity.kt
@@ -58,8 +58,11 @@ import com.r0adkll.deckbuilder.internal.analytics.Analytics
 import com.r0adkll.deckbuilder.internal.analytics.Event
 import com.r0adkll.deckbuilder.util.CardUtils
 import com.r0adkll.deckbuilder.util.MarketplaceHelper
+import com.r0adkll.deckbuilder.util.glide.ImageType
 import com.r0adkll.deckbuilder.util.extensions.drawable
 import com.r0adkll.deckbuilder.util.extensions.formatPrice
+import com.r0adkll.deckbuilder.util.extensions.loadOfflineUri
+import com.r0adkll.deckbuilder.util.extensions.loadPokemonCard
 import io.reactivex.Observable
 import kotlinx.android.synthetic.main.activity_card_detail.*
 import kotlinx.android.synthetic.main.layout_card_details.*
@@ -397,7 +400,7 @@ class CardDetailActivity : BaseActivity(), CardDetailUi, CardDetailUi.Intentions
             emptyView.isVisible = true
             emptyView.state = EmptyView.State.LOADING
             var request = GlideApp.with(this)
-                .load(card.imageUrlHiRes)
+                .loadPokemonCard(this, card, ImageType.HI_RES)
                 .transition(withCrossFade())
                 .listener(object : RequestListener<Drawable> {
                     override fun onLoadFailed(
@@ -431,7 +434,7 @@ class CardDetailActivity : BaseActivity(), CardDetailUi, CardDetailUi.Intentions
 
             // Load expansion symbol
             GlideApp.with(this)
-                .load(card.expansion?.symbolUrl)
+                .loadOfflineUri(this, card.expansion?.symbolUrl)
                 .transition(withCrossFade())
                 .into(expansionSymbol)
         }