+++ app/src/main/java/com/r0adkll/deckbuilder/arch/ui/features/decks/adapter/UiViewHolder.kt
@@ -30,6 +30,8 @@ import com.r0adkll.deckbuilder.arch.ui.features.decks.adapter.UiViewHolder.ViewT
 import com.r0adkll.deckbuilder.arch.ui.features.decks.adapter.UiViewHolder.ViewType.PREVIEW
 import com.r0adkll.deckbuilder.arch.ui.features.decks.adapter.UiViewHolder.ViewType.QUICK_START
 import com.r0adkll.deckbuilder.arch.ui.widgets.DeckImageView
+import com.r0adkll.deckbuilder.util.glide.ImageType
+import com.r0adkll.deckbuilder.util.extensions.loadPokemonCard
 import com.r0adkll.deckbuilder.util.stack
 import io.reactivex.disposables.CompositeDisposable
 import io.reactivex.disposables.Disposable
@@ -181,7 +183,7 @@ sealed class UiViewHolder<in I : Item>(itemView: View) : ViewHolder(itemView), D
                 }
             } ?: mostProminentCard(deck.cards)?.let {
                 GlideApp.with(itemView)
-                    .load(it.imageUrl)
+                    .loadPokemonCard(itemView.context, it, ImageType.NORMAL)
                     .placeholder(R.drawable.pokemon_card_back)
                     .into(image)
             }