+++ app/src/main/java/com/r0adkll/deckbuilder/arch/ui/features/testing/DeckTestingActivity.kt
@@ -39,8 +39,10 @@ import com.r0adkll.deckbuilder.arch.ui.features.testing.di.DeckTestingModule
 import com.r0adkll.deckbuilder.arch.ui.widgets.PokemonCardView
 import com.r0adkll.deckbuilder.internal.analytics.Analytics
 import com.r0adkll.deckbuilder.internal.analytics.Event
+import com.r0adkll.deckbuilder.util.glide.ImageType
 import com.r0adkll.deckbuilder.util.extensions.fromHtml
 import com.r0adkll.deckbuilder.util.extensions.isMulligan
+import com.r0adkll.deckbuilder.util.extensions.loadPokemonCard
 import io.pokemontcg.model.SubType
 import io.pokemontcg.model.SuperType
 import io.reactivex.Observable
@@ -178,7 +180,7 @@ class DeckTestingActivity : BaseActivity(), DeckTestingUi, DeckTestingUi.Intenti
     override fun showTestHand(hand: List<PokemonCard>) {
         val futures = hand.map {
             GlideApp.with(this)
-                .load(it.imageUrl)
+                .loadPokemonCard(this, it, ImageType.NORMAL)
                 .submit()
         }