+++ app/src/main/java/com/r0adkll/deckbuilder/arch/ui/features/testing/adapter/TestResultViewHolder.kt
@@ -11,6 +11,8 @@ import com.ftinc.kit.arch.util.bindView
 import com.r0adkll.deckbuilder.GlideApp
 import com.r0adkll.deckbuilder.R
 import com.r0adkll.deckbuilder.arch.ui.widgets.TestResultProgressView
+import com.r0adkll.deckbuilder.util.glide.ImageType
+import com.r0adkll.deckbuilder.util.extensions.loadPokemonCard
 
 @Suppress("MagicNumber")
 class TestResultViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
@@ -27,9 +29,11 @@ class TestResultViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
         background.requestLayout()
         percent.text = "${background.percentage.toInt()}%"
 
-        GlideApp.with(itemView)
-            .load(item.pokemonCard?.imageUrl)
-            .into(imageView)
+        if (item.pokemonCard != null) {
+            GlideApp.with(itemView)
+                .loadPokemonCard(itemView.context, item.pokemonCard, ImageType.NORMAL)
+                .into(imageView)
+        }
     }
 
     companion object {