+++ app/src/main/java/com/r0adkll/deckbuilder/internal/di/AppComponent.kt
@@ -29,6 +29,8 @@ import com.r0adkll.deckbuilder.arch.ui.features.search.di.SearchComponent
 import com.r0adkll.deckbuilder.arch.ui.features.setbrowser.di.SetBrowserComponent
 import com.r0adkll.deckbuilder.arch.ui.features.setbrowser.di.SetBrowserModule
 import com.r0adkll.deckbuilder.arch.ui.features.settings.SettingsActivity
+import com.r0adkll.deckbuilder.arch.ui.features.settings.cache.di.ManageCacheComponent
+import com.r0adkll.deckbuilder.arch.ui.features.settings.cache.di.ManageCacheModule
 import com.r0adkll.deckbuilder.arch.ui.features.setup.SetupActivity
 import com.r0adkll.deckbuilder.arch.ui.features.testing.di.DeckTestingComponent
 import com.r0adkll.deckbuilder.arch.ui.features.testing.di.DeckTestingModule
@@ -62,6 +64,7 @@ interface AppComponent {
     fun plus(module: DeckImageModule): DeckImageComponent
     fun plus(module: DeckTestingModule): DeckTestingComponent
     fun plus(module: CollectionSetModule): CollectionSetComponent
+    fun plus(module: ManageCacheModule): ManageCacheComponent
 
     fun searchComponentBuilder(): SearchComponent.Builder
     fun deckBuilderComponentBuilder(): DeckBuilderComponent.Builder